requirements.txt
aiosmtpd
channels
channels-redis
chardet
defusedxml
dj-database-url
django
django-bootstrap3
django-filter
django-pure-pagination
djangorestframework
dnslib
docker
flanker
gevent
gunicorn
hashids
huey
ipython
jsonfield
pipenv-to-requirements
psycopg2-binary
python-dotenv
redis
requests
requests-oauthlib
tenacity
asgiref